+++ spec/rubocop/cop/style/arguments_forwarding_spec.rb
@@ -219,8 +219,24 @@ RSpec.describe RuboCop::Cop::Style::ArgumentsForwarding, :config do
 
     it 'does not register an offense when different argument names are used' do
       expect_no_offenses(<<~RUBY)
+        def foo(arg)
+          bar(argument)
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when different splat argument names are used' do
+      expect_no_offenses(<<~RUBY)
         def foo(*args, &block)
           bar(*arguments, &block)
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when different kwrest argument names are used' do
+      expect_no_offenses(<<~RUBY)
+        def foo(**kwargs, &block)
+          bar(**kwarguments, &block)
         end
       RUBY
     end
@@ -303,6 +319,15 @@ RSpec.describe RuboCop::Cop::Style::ArgumentsForwarding, :config do
           bar(*args, **kwargs, &block)
           bar(*args, &block)
           bar(**kwargs, &block)
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when always forwarding the block but not other args' do
+      expect_no_offenses(<<~RUBY)
+        def foo(*args, &block)
+          bar(*args, &block)
+          bar(&block)
         end
       RUBY
     end
@@ -1195,6 +1220,15 @@ RSpec.describe RuboCop::Cop::Style::ArgumentsForwarding, :config do
       RUBY
     end
 
+    it 'does not register an offense when always forwarding the block but not other args' do
+      expect_no_offenses(<<~RUBY)
+        def foo(*, &block)
+          bar(*, &block)
+          bar(&block)
+        end
+      RUBY
+    end
+
     it 'registers an offense when args are forwarded at several call sites' do
       expect_offense(<<~RUBY)
         def foo(bar, *args)
@@ -1514,21 +1548,53 @@ RSpec.describe RuboCop::Cop::Style::ArgumentsForwarding, :config do
       RUBY
     end
 
-    it 'registers an offense when forwarding to a method in block' do
-      expect_offense(<<~RUBY)
+    it 'does not register an offense when rest arguments forwarding to a method in block' do
+      expect_no_offenses(<<~RUBY)
         def foo(*args, &block)
-                ^^^^^ Use anonymous positional arguments forwarding (`*`).
           do_something do
             bar(*args, &block)
-                ^^^^^ Use anonymous positional arguments forwarding (`*`).
           end
         end
       RUBY
+    end
 
-      expect_correction(<<~RUBY)
-        def foo(*, &block)
+    it 'does not register an offense when rest arguments forwarding to a method in numbered block' do
+      expect_no_offenses(<<~RUBY)
+        def foo(*args, &block)
           do_something do
-            bar(*, &block)
+            bar(*args, &block)
+            baz(_1)
+          end
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when keyword rest arguments forwarding to a method in block' do
+      expect_no_offenses(<<~RUBY)
+        def foo(**kwargs, &block)
+          do_something do
+            bar(**kwargs, &block)
+          end
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when keyword rest arguments forwarding to a method in numbered block' do
+      expect_no_offenses(<<~RUBY)
+        def foo(**kwargs, &block)
+          do_something do
+            bar(**kwargs, &block)
+            baz(_1)
+          end
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when rest arguments and keyword rest arguments forwarding to a method in block' do
+      expect_no_offenses(<<~RUBY)
+        def foo(*args, **kwargs)
+          block_method do
+            bar(*args, **kwargs)
           end
         end
       RUBY
@@ -1592,8 +1658,24 @@ RSpec.describe RuboCop::Cop::Style::ArgumentsForwarding, :config do
 
     it 'does not register an offense when different argument names are used' do
       expect_no_offenses(<<~RUBY)
+        def foo(arg)
+          bar(argument)
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when different splat argument names are used' do
+      expect_no_offenses(<<~RUBY)
         def foo(*args, &block)
           bar(*arguments, &block)
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when different kwrest argument names are used' do
+      expect_no_offenses(<<~RUBY)
+        def foo(**kwargs, &block)
+          bar(**kwarguments, &block)
         end
       RUBY
     end