+++ spec/rubocop/cop/style/hash_each_methods_spec.rb
@@ -81,12 +81,50 @@ RSpec.describe RuboCop::Cop::Style::HashEachMethods, :config do
         expect_no_offenses('foo.each { |(_, k), v| do_something(k, v) }')
       end
 
+      it 'does not register an offense when the destructed rest value block arguments of `Enumerable#each` method are used' do
+        expect_no_offenses('foo.each { |k, (_, *v)| do_something(k, *v) }')
+      end
+
+      it 'does not register an offense when the destructed rest key block arguments of `Enumerable#each` method are used' do
+        expect_no_offenses('foo.each { |(_, *k), v| do_something(*k, v) }')
+      end
+
       it 'does not register an offense when the single block argument of `Enumerable#each` method is used' do
         expect_no_offenses('foo.each { |e| do_something(e) }')
       end
 
       it 'does not register an offense when the parenthesized key and value block arguments of `Enumerable#each` method are unused' do
         expect_no_offenses('foo.each { |(k, v)| do_something(e) }')
+      end
+
+      it 'does not register an offense when the rest value block argument of `Enumerable#each` method is used' do
+        expect_no_offenses('foo.each { |k, *v| do_something(k, *v) }')
+      end
+
+      it 'does not register an offense when the rest key block argument of `Enumerable#each` method is used' do
+        expect_no_offenses('foo.each { |*k, v| do_something(*k, v) }')
+      end
+
+      it 'registers an offense when the rest value block argument of `Enumerable#each` method is unused' do
+        expect_offense(<<~RUBY)
+          foo.each { |k, *v| do_something(*v) }
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Use `each_value` instead of `each` and remove the unused `k` block argument.
+        RUBY
+
+        expect_correction(<<~RUBY)
+          foo.each_value { |*v| do_something(*v) }
+        RUBY
+      end
+
+      it 'registers an offense when the rest key block argument of `Enumerable#each` method is unused' do
+        expect_offense(<<~RUBY)
+          foo.each { |*k, v| do_something(*k) }
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Use `each_key` instead of `each` and remove the unused `v` block argument.
+        RUBY
+
+        expect_correction(<<~RUBY)
+          foo.each_key { |*k| do_something(*k) }
+        RUBY
       end
 
       it 'registers an offense when the value block argument of `Enumerable#each` method is unused' do
@@ -141,6 +179,23 @@ RSpec.describe RuboCop::Cop::Style::HashEachMethods, :config do
 
         expect_correction(<<~RUBY)
           foo.each_value { |v| do_something(v) }
+        RUBY
+      end
+
+      it 'registers an offense and corrects when `{hash: :literal}.keys.each`' do
+        expect_offense(<<~RUBY)
+          {hash: :literal}.keys.each { |k| p k }
+                           ^^^^^^^^^ Use `each_key` instead of `keys.each`.
+        RUBY
+
+        expect_correction(<<~RUBY)
+          {hash: :literal}.each_key { |k| p k }
+        RUBY
+      end
+
+      it 'does not register an offense when `[[1, 2, 3], [4 ,5, 6]].each`' do
+        expect_no_offenses(<<~RUBY)
+          [[1, 2, 3], [4, 5, 6]].each { |a, _| p a }
         RUBY
       end