+++ spec/rubocop/cop/style/redundant_parentheses_spec.rb
@@ -508,6 +508,36 @@ RSpec.describe RuboCop::Cop::Style::RedundantParentheses, :config do
     RUBY
   end
 
+  it 'registers an offense when the use of parentheses around `&&` expressions in assignment' do
+    expect_offense(<<~RUBY)
+      var = (foo && bar)
+            ^^^^^^^^^^^^ Don't use parentheses around a logical expression.
+    RUBY
+
+    expect_correction(<<~RUBY)
+      var = foo && bar
+    RUBY
+  end
+
+  it 'registers an offense when the use of parentheses around `||` expressions in assignment' do
+    expect_offense(<<~RUBY)
+      var = (foo || bar)
+            ^^^^^^^^^^^^ Don't use parentheses around a logical expression.
+    RUBY
+
+    expect_correction(<<~RUBY)
+      var = foo || bar
+    RUBY
+  end
+
+  it 'accepts the use of parentheses around `or` expressions in assignment' do
+    expect_no_offenses('var = (foo or bar)')
+  end
+
+  it 'accepts the use of parentheses around `and` expressions in assignment' do
+    expect_no_offenses('var = (foo and bar)')
+  end
+
   it 'accepts parentheses around a method call with unparenthesized arguments' do
     expect_no_offenses('(a 1, 2) && (1 + 1)')
   end
@@ -526,6 +556,30 @@ RSpec.describe RuboCop::Cop::Style::RedundantParentheses, :config do
 
   it 'accepts parentheses around an erange' do
     expect_no_offenses('(a...b)')
+  end
+
+  it 'accepts an irange starting is a parenthesized condition' do
+    expect_no_offenses('(a || b)..c')
+  end
+
+  it 'accepts an erange starting is a parenthesized condition' do
+    expect_no_offenses('(a || b)...c')
+  end
+
+  it 'accepts an irange ending is a parenthesized condition' do
+    expect_no_offenses('a..(b || c)')
+  end
+
+  it 'accepts an erange ending is a parenthesized condition' do
+    expect_no_offenses('a...(b || c)')
+  end
+
+  it 'accepts regexp literal attempts to match against a parenthesized condition' do
+    expect_no_offenses('/regexp/ =~ (b || c)')
+  end
+
+  it 'accepts variable attempts to match against a parenthesized condition' do
+    expect_no_offenses('regexp =~ (b || c)')
   end
 
   it 'registers parentheses around `||` logical operator keywords in method definition' do
@@ -833,5 +887,110 @@ RSpec.describe RuboCop::Cop::Style::RedundantParentheses, :config do
     it_behaves_like 'allowed parentheses', '{ baz: 2 }', 'a hash literal'
     it_behaves_like 'allowed parentheses', '1..2', 'a range literal'
     it_behaves_like 'allowed parentheses', '1', 'an int literal'
+  end
+
+  context 'when `AllowInMultilineConditions: true` of `Style/ParenthesesAroundCondition`' do
+    let(:other_cops) do
+      {
+        'Style/ParenthesesAroundCondition' => {
+          'Enabled' => enabled, 'AllowInMultilineConditions' => true
+        }
+      }
+    end
+
+    context 'when `Style/ParenthesesAroundCondition` is enabled' do
+      let(:enabled) { true }
+
+      context 'when single line conditions' do
+        it_behaves_like 'redundant', '(x && y)', 'x && y', 'a logical expression'
+        it_behaves_like 'redundant', '(x || y)', 'x || y', 'a logical expression'
+        it_behaves_like 'redundant', '(x and y)', 'x and y', 'a logical expression'
+        it_behaves_like 'redundant', '(x or y)', 'x or y', 'a logical expression'
+      end
+
+      context 'when multiline conditions' do
+        it_behaves_like 'plausible', <<~RUBY
+          (x &&
+           y)
+        RUBY
+        it_behaves_like 'plausible', <<~RUBY
+          (x ||
+           y)
+        RUBY
+        it_behaves_like 'plausible', <<~RUBY
+          (x and
+           y)
+        RUBY
+        it_behaves_like 'plausible', <<~RUBY
+          (x or
+           y)
+        RUBY
+      end
+    end
+
+    context 'when `Style/ParenthesesAroundCondition` is disabled' do
+      let(:enabled) { false }
+
+      context 'when single line conditions' do
+        it_behaves_like 'redundant', '(x && y)', 'x && y', 'a logical expression'
+        it_behaves_like 'redundant', '(x || y)', 'x || y', 'a logical expression'
+        it_behaves_like 'redundant', '(x and y)', 'x and y', 'a logical expression'
+        it_behaves_like 'redundant', '(x or y)', 'x or y', 'a logical expression'
+      end
+
+      context 'when multiline conditions' do
+        it 'registers an offense when using `&&`' do
+          expect_offense(<<~RUBY)
+            (x &&
+            ^^^^^ Don't use parentheses around a logical expression.
+             y)
+          RUBY
+
+          expect_correction(<<~RUBY)
+            x &&
+             y
+          RUBY
+        end
+
+        it 'registers an offense when using `||`' do
+          expect_offense(<<~RUBY)
+            (x ||
+            ^^^^^ Don't use parentheses around a logical expression.
+             y)
+          RUBY
+
+          expect_correction(<<~RUBY)
+            x ||
+             y
+          RUBY
+        end
+
+        it 'registers an offense when using `and`' do
+          expect_offense(<<~RUBY)
+            (x and
+            ^^^^^^ Don't use parentheses around a logical expression.
+             y)
+          RUBY
+
+          expect_correction(<<~RUBY)
+            x and
+             y
+          RUBY
+        end
+
+        it 'registers an offense when using `or`' do
+          expect_offense(<<~RUBY)
+            (x or
+            ^^^^^ Don't use parentheses around a logical expression.
+             y)
+          RUBY
+
+          expect_correction(<<~RUBY)
+            x or
+             y
+          RUBY
+        end
+      end
+    end
   end
 end